backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI Integration
google-generativeai==0.3.1

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Security Scanners (Core)
bandit==1.7.5

# GitHub Integration
PyGithub==2.1.1
cryptography==41.0.7

# Async & Caching
redis==5.0.1

# Logging
loguru==0.7.2

# Data Export
pandas==2.1.4
openpyxl==3.1.2

# Utilities
httpx==0.25.2
python-multipart==0.0.6
